Rename directory and component
