+++ src/app/shared/components/terminal/style.scss
