+++ src/app/shared/components/terminal/template.html
