+++ src/app/shared/components/terminal/terminal-status-bar/component.ts
