+++ src/app/shared/components/terminal/terminal-status-bar/style.scss
