+++ src/app/shared/components/terminal/terminal-status-bar/template.html
