+++ src/app/shared/components/terminal/terminal-status-bar/theme.scss
