+++ src/app/shared/components/terminal/terminal-toolbar/component.ts
