+++ src/app/shared/components/terminal/terminal-toolbar/style.scss
