+++ src/app/shared/components/terminal/terminal-toolbar/template.html
