+++ src/app/shared/components/terminal/terminal-toolbar/theme.scss
