+++ src/assets/css/theme/_component.scss
@@ -37,8 +37,8 @@
 @import 'src/app/shared/components/chip/theme';
 @import 'src/app/shared/components/tab-card/theme';
 @import 'src/app/shared/components/openstack-credentials/theme';
-@import 'src/app/shared/components/xterm-terminal/terminal-toolbar/theme';
-@import 'src/app/shared/components/xterm-terminal/terminal-status-bar/theme';
+@import 'src/app/shared/components/terminal/terminal-toolbar/theme';
+@import 'src/app/shared/components/terminal/terminal-status-bar/theme';
 @import 'src/app/cluster/details/cluster/overlay-terminal/theme';
 @import 'src/app/cluster/details/cluster/rbac/add-binding/theme';
 @import 'src/app/settings/admin/bucket-settings/destinations/theme';
